Hai Nam: Fix vnpay

+++ src/main/java/com/hqt/happyhostel/servlet/PaymentServlets/VNPayPaymentServlet.java
@@ -37,7 +37,7 @@ public class VNPayPaymentServlet extends HttpServlet {
             String vnp_OrderInfo = bill.getBillTitle();
             String vnp_IpAddr = Config.getIpAddress(req);
             String vnp_TmnCode = Config.vnp_TmnCode;
-            int amount = bill.getTotalMoney() * 100;
+            long amount = (long)bill.getTotalMoney() * 100;
             Map<String, String> vnp_Params = new HashMap<>();
             vnp_Params.put("vnp_Version", vnp_Version);
             vnp_Params.put("vnp_Command", vnp_Command);
